requirements.txt
Flask==2.3.2
scikit-learn==1.6.0
joblib==1.3.2
numpy==1.23.5
Colorama==0.4.6
mysql-connector-python==8.0.33  # Add MySQL connector